scx_userland: never set CPU ownership to the user-scheduler pid
Always consider the CPU where the user-space scheduler is running as
idle.

Considering the CPU where the user-space scheduler is running as busy
doesn't really provide any benefit, since the user-space scheduler is
constantly dispatching an amount of tasks equal to the amount of idle
CPUs and then yields (therefore its own CPU should be considered idle).

This also allows to reduce the user-space scheduler CPU usage,
especially when the system is mostly idle.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>