scx_rustland: re-introduce per-CPU DSQ + a global shared DSQ
With commit c6ada25 ("scx_rustland: use custom pcpu DSQ instead of
SCX_DSQ_LOCAL{_ON}") we tried to introduce custom per-CPU DSQs, instead
of using SCX_DSQ_LOCAL and SCX_DSQ_LOCAL_ON to dispatch tasks.

This was required, because dispatching tasks using SCX_DSQ_LOCAL_ON
doesn't provide a guarantee that the cpumask, checked at dispatch time
to determine the validity of a target CPU, remains valid.

This method solved the cpumask validity issue, but unfortunately it
introduced a noticeable performance regression and a potential
starvation issue (that were probably caused by the same problem): if a
task is assigned to a CPU in select_cpu() and the scheduler decides to
dispatch it on a different CPU, the task will be added to the new CPU's
DSQ, but if no dispatch event happens there, the task may remain stuck
in the per-CPU DSQ for a long time, triggering the sched-ext watchdog
timeout that would kick out the scheduler, for example:

  12:53:28 [WARN] FAIL: IPC:CSteamEngin[7217] failed to run for 6.482s (err=1026)
  12:53:28 [INFO] Unregister RustLand scheduler

Therefore, we reverted this change with 6d89ece ("scx_rustland: dispatch
tasks only on the global DSQ"), dispatching all the tasks to the global
DSQ, completely delegating the kernel to distribute tasks among the
available CPUs.

This is not the ideal solution, because we still want to give the
possibility to the user-space scheduler to assign tasks to specific
CPUs.

Therefore, re-introduce distinct per-CPU DSQs, but also provide a global
shared DSQ. Tasks dispatched in the per-CPU DSQs are consumed from the
dispatch() callback of their corresponding CPU, tasks dispatched in the
global shared DSQ are consumed from any CPU.

In this way the per-CPU DSQs can be used to dispatch tasks to specific
CPUs and, if a CPU is not specified, tasks will send to the shared DSQ
and consumed on the first CPU available.

Apparently all this logic works only if tasks are dispatched to the DSQ
associated to the CPU selected in the select_cpu() callback, or if they
are dispatched to the shared DSQ.

If a task is dispatched on a CPU different than the one determined in
select_cpu() we may still hit the starvation issue.

Therefore, as a temporary workaround, when this case happens, simply
redirect the task to the global shared DSQ.

This is definitely not ideal, since we may want to honor the decision
made by the user-space scheduler that should have the ability to
override the decision made the built-in selection logic (or by any
"default" logic implemented in select_cpu()).

But, until we figure out a proper solution, this method seems to work
and it makes the scheduler stable and responsive.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>