scx_user: move scx_rustland backend code to scx_user
Move the BPF component of scx_rustland to scx_user and make it available
to other user-space schedulers.

NOTE: main.bpf.c and bpf.rs are not pre-compiled in the scx_user crate,
they need to be included in the user-space scheduler's source code in
order to be compiled/linked properly.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>