scx_rustland: update documentation
scx_rustland has significantly evolved since its original design.

With the introduction of scx_user and the inclusion of the scx_fifo
example, scx_rustland's focus can be shifted from solely being an
"easy-to-read Rust scheduler template" to a fully functional scheduler.

For this reason, update the README and documentation to reflect its
revised design, objectives, and intended use cases.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>