scx_rustland_core: use a BPF_MAP_TYPE_USER_RINGBUF for dispatched tasks
Replace the BPF_MAP_TYPE_QUEUE with a BPF_MAP_TYPE_USER_RINGBUF to store
the tasks dispatched from the user-space component.

This allows to not use the bpf() syscalls when tasks are dispatched from
user-space to the BPF counterpart, reducing the scheduling overhead and
providing a significant performance improvement in system throughput.

According to the experimental results this change seems to reduce the
scheduling overhead of around 30-35% when the system is over
commissioned.

Link: https://github.com/libbpf/libbpf-rs/pull/776
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>