scx_bpfland: consider wakeups by per-CPU kthreads as synchronous
If a per-CPU kthread runs on the wakee's previous CPU, treat the run of
the wakee as a sync wakeup.

A typical example of this behavior is IO completion: a task had queued
work for the per-CPU kthread, then the kthread has finished and it can
give control back to the task.

Signed-off-by: Andrea Righi <arighi@nvidia.com>