scx_bpfland: boost per-CPU tasks
Tasks that can run on a single CPU can be penalized more than others,
because they don't have the chance to migrate on another CPU if their
only usable CPU is busy.

To compensate this always assign their deadline as zero and schedule
them purely in function of their vruntime.

Signed-off-by: Andrea Righi <arighi@nvidia.com>