scx_rustland_core: Detect dequeue events
When a task is sitting in the user-space scheduler it may be dequeued by
the core scheduler to modify properties (i.e., CPU affinity, priority,
etc.) and then re-enqueued again.

The user-space scheduler, however, has no visibility of these underlying
changes. This can lead to spurious enqueues that the sched_ext core
correctly ignores, but they can still generate unnecessary overhead and
increase locking contention on the DSQs.

To address this, store an enqueue generation counter in the task context
sent to the user-space scheduler. Then, when a task is dispatched, we
can check the generation counter to determine whether the task was
re-enqueued or not. If it was, the current enqueue event can be ignored,
avoiding an unnecessary DSQ insert operation.

This mechanism reduces overhead and helps minimize DSQ locking
contention.

Signed-off-by: Andrea Righi <arighi@nvidia.com>